+++ internal/pkg/instrumentation/bpf/database/sql/bpf_x86_bpfel.go
@@ -64,9 +64,10 @@ func loadBpfObjects(obj interface{}, opts *ebpf.CollectionOptions) error {
 type bpfSpecs struct {
 	bpfProgramSpecs
 	bpfMapSpecs
+	bpfVariableSpecs
 }
 
-// bpfSpecs contains programs before they are loaded into the kernel.
+// bpfProgramSpecs contains programs before they are loaded into the kernel.
 //
 // It can be passed ebpf.CollectionSpec.Assign.
 type bpfProgramSpecs struct {
@@ -90,12 +91,25 @@ type bpfMapSpecs struct {
 	TrackedSpansBySc      *ebpf.MapSpec `ebpf:"tracked_spans_by_sc"`
 }
 
+// bpfVariableSpecs contains global variables before they are loaded into the kernel.
+//
+// It can be passed ebpf.CollectionSpec.Assign.
+type bpfVariableSpecs struct {
+	EndAddr                  *ebpf.VariableSpec `ebpf:"end_addr"`
+	Hex                      *ebpf.VariableSpec `ebpf:"hex"`
+	IsRegistersAbi           *ebpf.VariableSpec `ebpf:"is_registers_abi"`
+	ShouldIncludeDbStatement *ebpf.VariableSpec `ebpf:"should_include_db_statement"`
+	StartAddr                *ebpf.VariableSpec `ebpf:"start_addr"`
+	TotalCpus                *ebpf.VariableSpec `ebpf:"total_cpus"`
+}
+
 // bpfObjects contains all objects after they have been loaded into the kernel.
 //
 // It can be passed to loadBpfObjects or ebpf.CollectionSpec.LoadAndAssign.
 type bpfObjects struct {
 	bpfPrograms
 	bpfMaps
+	bpfVariables
 }
 
 func (o *bpfObjects) Close() error {
@@ -130,6 +144,18 @@ func (m *bpfMaps) Close() error {
 		m.SqlEvents,
 		m.TrackedSpansBySc,
 	)
+}
+
+// bpfVariables contains all global variables after they have been loaded into the kernel.
+//
+// It can be passed to loadBpfObjects or ebpf.CollectionSpec.LoadAndAssign.
+type bpfVariables struct {
+	EndAddr                  *ebpf.Variable `ebpf:"end_addr"`
+	Hex                      *ebpf.Variable `ebpf:"hex"`
+	IsRegistersAbi           *ebpf.Variable `ebpf:"is_registers_abi"`
+	ShouldIncludeDbStatement *ebpf.Variable `ebpf:"should_include_db_statement"`
+	StartAddr                *ebpf.Variable `ebpf:"start_addr"`
+	TotalCpus                *ebpf.Variable `ebpf:"total_cpus"`
 }
 
 // bpfPrograms contains all programs after they have been loaded into the kernel.